MODIFY: file structure changed and ssim working
