+++ 11-centre_clipped.py
