+++ front_end_schemes.py
@@ -7,7 +7,43 @@ from cued_sf2_lab.dct import colxfm, dct_ii, regroup
 from cued_sf2_lab.lbt import pot_ii
 from useful_functions import *
 
-# If want to generate DCT (equal RMS), set s=None
+# Forward and inverse DCT/LBT, set s=None to generate DCT
+def forward_dct_lbt(X, C, s=None, rise1_ratio=0.5, supp_comp_num=0):
+    """
+    Note C is the DCT matrix of coefficients, generated using dct_ii(N).
+    C is input paramter instead of block_size to avoid having to re-compute C in every function when one function calls another.
+    When s = None we perform the DCT instead of the LBT.
+    """
+    if s == None:
+        Xp = X
+    else:
+        N = C.shape[0]
+        Pf, Pr = pot_ii(N, s)
+        t = np.s_[N//2:-N//2]
+        Xp = X.copy()
+        Xp[t, :] = colxfm(Xp[t, :], Pf)
+        Xp[:, t] = colxfm(Xp[:, t].T, Pf).T
+
+    Y = colxfm(colxfm(Xp, C).T, C).T
+    Y = suppress_components(Y, C.shape[0], supp_comp_num)
+
+    return Y
+
+def inverse_dct_lbt(Y, C, s=None, rise1_ratio=0.5, supp_comp_num=0):
+    Z = colxfm(colxfm(Y.T, C.T).T, C.T)
+    
+    if s == None:
+        return Z
+    else:
+        N = C.shape[0]
+        Pf, Pr = pot_ii(N, s)
+        t = np.s_[N//2:-N//2]
+        Zp = Z.copy()
+        Zp[:, t] = colxfm(Zp[:, t].T, Pr.T).T
+        Zp[t, :] = colxfm(Zp[t, :], Pr.T)
+        return Zp
+
+# Equal rms DCT/LBT. Returns compression ratio, Z
 def gen_dct_lbt_equal_rms(X, block_size, s=None, rise1_ratio=0.5, supp_comp_num=0):
     if block_size <= 0 or (block_size & (block_size - 1)) != 0:
         return "Error the block size is not a power of two!"
@@ -32,6 +68,26 @@ def gen_dct_lbt_equal_rms(X, block_size, s=None, rise1_ratio=0.5, supp_comp_num=
         Zp[t,:] = colxfm(Zp[t,:], Pr.T)
         return comp_ratio, Zp
 
+# Forward and inverse DWT
+def nlevdwt(X, n):
+    Y = X.copy()
+    m = X.shape[0]
+    for i in range(n):
+        Y[:m,:m] = dwt(Y[:m,:m])
+        m //= 2
+    
+    return Y
+
+def nlevidwt(Y, n):
+    m = Y.shape[0] // (2 ** (n - 1))
+    Z = Y.copy()
+    for i in range(n):
+        Z[:m,:m] = idwt(Z[:m,:m])
+        m = m * 2
+
+    return Z
+
+# Equal MSE DWT. Returns compression ratio, Z
 def gen_dwt_equal_mse(X, num_levels):
     energies = impulse_energies(num_levels)
 
@@ -54,16 +110,3 @@ def gen_dwt_equal_mse(X, num_levels):
     Z = nlevidwt(Yq, num_levels)
 
     return comp_ratio, Z
-
-
-def eightDCTjpeg(X):
-    step_size = 16
-    calc_bits = 100000
-    while calc_bits > 40960:
-        vlctemp, _ = jpegenc(X, step_size)
-        calc_bits = vlctemp[:,1].sum()
-        step_size += 1
-    vlc, _ = jpegenc(X, step_size)
-    Z = jpegdec(vlc, step_size)
-    rmserr = np.std(Z - X)
-    return Z, rmserr, calc_bits
